renaming function class file
