+++ src/function_classes/chebyshev.py
